Rename package to org.hypertrace.attributeutils
