+++ attribute-utils/src/main/java/org/hypertrace/attributeutils/db/OTelDbAttributes.java
@@ -1,4 +1,4 @@
-package org.hypertrace.attribute.db;
+package org.hypertrace.attributeutils.db;
 
 /**
  * OTEL specific attributes for database
